load-default-config action
